+++ tests/integration/weaviate_memory_tests.py
@@ -1,14 +1,11 @@
-import os
-import sys
 import unittest
-from unittest import mock
 from uuid import uuid4
 
 from weaviate import Client
 from weaviate.util import get_valid_uuid
 
 from autogpt.config import Config
-from autogpt.memory.base import get_ada_embedding
+from autogpt.llm_utils import get_ada_embedding
 from autogpt.memory.weaviate import WeaviateMemory
